+++ pdt_client/commands.py
@@ -281,20 +281,21 @@ def get_case_revisions(url, username, password, ci_project, release, case=None):
     try:
         response.raise_for_status()
         for case in response.json():
-            print('{0}\t{1}'.format(case['id'], case['revision']))
+            print('{case[id]}\t{case[revision]}\t{case[title]}'.format(case=case))
     except Exception:
         pprint.pprint(response.json())
         raise
 
 
-def deploy(url, username, password, instance, ci_project, release, status, log, cases):
+def deploy(url, username, password, instance, ci_project, release, status, log, cases, revision):
     """Report the deployment."""
     data = dict(
         status=status,
         instance=dict(name=instance, ci_project=dict(name=ci_project)),
         release=dict(number=release),
         cases=[dict(id=case) for case in cases],
-        log=log.read()
+        log=log.read(),
+        **dict(revision=revision) if revision else {}
     )
     response = requests.post(
         '{url}/api/deployment-reports/'.format(url=url),